refactor: extract web serial bindings
