+++ packages/bindings-browser/package.json
@@ -15,6 +15,10 @@
     "./fs": {
       "@@dev": "./src/fs/browser.ts",
       "default": "./build/fs/browser.js"
+    },
+    "./serial": {
+      "@@dev": "./src/serial/browser.ts",
+      "default": "./build/serial/browser.js"
     }
   },
   "files": [
@@ -45,6 +49,8 @@
     "clean": "del-cli build/ \"*.tsbuildinfo\""
   },
   "devDependencies": {
+    "@types/w3c-web-serial": "^1.0.7",
+    "@zwave-js/serial": "workspace:*",
     "@zwave-js/shared": "workspace:*",
     "del-cli": "^6.0.0",
     "typescript": "5.7.3",
